// lib/types/enums/waf.ts
export enum WebAclScope {
    CLOUDFRONT = "CLOUDFRONT",
    REGIONAL = "REGIONAL"
  }
// ...
export enum AwsManagedRules {
    COMMON_RULE_SET = "AWSManagedRulesCommonRuleSet",
    ADMIN_PROTECTION_RULE_SET = "AWSManagedRulesAdminProtectionRuleSet",
    KNOWN_BAD_INPUTS_RULE_SET = "AWSManagedRulesKnownBadInputsRuleSet",
    SQLI_RULE_SET = "AWSManagedRulesSQLiRuleSet",
    LINUX_RULE_SET = "AWSManagedRulesLinuxRuleSet",
    UNIX_RULE_SET = "AWSManagedRulesUnixRuleSet",
    WINDOWS_RULE_SET = "AWSManagedRulesWindowsRuleSet",
    PHP_RULE_SET = "AWSManagedRulesPHPRuleSet",
    WORDPRESS_RULE_SET = "AWSManagedRulesWordPressRuleSet",
    AMAZON_IP_REPUTATION_LIST = "AWSManagedRulesAmazonIpReputationList",
    ANONYMOUS_IP_LIST = "AWSManagedRulesAnonymousIpList",
    BOT_CONTROL_RULE_SET = "AWSManagedRulesBotControlRuleSet",
    ATP_RULE_SET = "AWSManagedRulesATPRuleSet",
    ACFP_RULE_SET = "AWSManagedRulesACFPRuleSet"
  }
// ...
export enum ManagedRuleGroupVendor {
    AWS = "AWS"
  }
// ...
export enum WebAclTypeEnum {
    ELASTICLOADBALANCINGV2_LOADBALANCER = "AWS::ElasticLoadBalancingV2::LoadBalancer",
    CLOUDFRONT_DISTRIBUTION = "AWS::CloudFront::Distribution",
    APIGATEWAYV2_API = "AWS::ApiGatewayV2::Api",
    APIGATEWAY_STAGE = "AWS::ApiGateway::Stage",
    COGNITO_USERPOOL = "AWS::Cognito::UserPool",
    APPSYNC_GRAPHQLAPI  = "AWS::AppSync::GraphQLApi"
  }